refactor: 어드민 페이지 쿠폰 관련 섹션 분리
